pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "passive-income-blog"
version = "0.1.0"
description = "Automated SEO blog generation and monetization system"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.0",
    "pytrends>=4.9.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
]

[project.scripts]
generate = "pipeline.main:main"
